+++ build.gradle
@@ -37,6 +37,9 @@ dependencies {
     // Lombok
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
+
+    testCompileOnly 'org.projectlombok:lombok'
+    testAnnotationProcessor 'org.projectlombok:lombok'
     
     // Test Dependencies
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
@@ -47,6 +50,7 @@ dependencies {
     // Cucumber
     testImplementation "io.cucumber:cucumber-java:${cucumberVersion}"
     testImplementation "io.cucumber:cucumber-junit-platform-engine:${cucumberVersion}"
+    testImplementation "io.cucumber:cucumber-picocontainer:${cucumberVersion}"
     
     // RestAssured
     testImplementation "io.rest-assured:rest-assured:${restAssuredVersion}"
